+++ CHANGELOG.md
@@ -6,6 +6,11 @@
     - Going Public
     - Updated README and TODO elements
         - Moved Issues to GitHub
+- Added the first test cases
+    - Many String Functions
+    - General Memory tests
+- Fixed Issues
+    - str_ends_with was not working correctly. It was fixed
 
 ## v0.3.0